refactor(tui): extract slash command entry component
